Removed old model files
The entire soruce tree (with exception for Stacky, a small module
that doesn't need to change) has been deleted to make way for a new
source tree based in part on the original code, and in part on the
new command register model.

The model is as follows (in its current form):
Command register functionality

Command register:  tuple values, one pointer to function, one string
                   classifying its command. Each command is then added
                   to the list of those the language will parse. This
                   will extend the extendability of the program

Command Parser:    Similar to what commandparser:: was supposed to be,
                   a simple parser which shunts work off to other
                   parts of the program. Does nothing more than just
                   call the correct function when a command is
                   entered.

Language Core:     Contains all core language variables and constants,
                   along with some convienience functions for
                   maintaining them. Essentially a souped-up defs.hh.
                   Functions like read() and todouble() will reside
                   here.

Modules interface: Where all modules should be created. A class in
                   and of itself, but built to call all other modules.
                   Where all current modules will reside in one form
                   or another.