Added new commandparser changes
The command parser now includes more headers for eventual language
features like the repetition buffer and script reading.

These changes were made in part to confirm that other changes would
allow the program to compile.

Also, the initilizer for commandparser requires a commandregister
object to be passed as a pointer. So instead of commandparser p;
the syntax commandregister r; commandparser p (&r); is correct.