Fixed compile errors, removed git string
The version string no longer includes commit strings.

Several functions with a declared type didn't return anything, the
statement was missing. This commit adds all the required returns (as
		caught by clang)

Changed indent to Emacs default in many of the files. This is to reflect
a personal editor switch from (+ kate kwrite gedit geany mousepad
		leafpad vim) to (+ emacs)